Updater.ini example file
The Updater.ini file is required by Updater.py.  It contains a comma separated list of the name of the program, the latest version, and the installation command string to silently install the updated version of the program.  These values must be compatible with how the program is listed in the in "uninstall" key in the registry.